Add msys2 include path for Qt4 into configure.ac
Using the variable QTDIR was not possible to find the Qt4 include
headers. The $QTDIR/include/qt4 was added to the list of possible
paths. This should work with default MSYS2 installation.
Setting QTDIR still work in case the user has MSYS2 installed
elsewhereto. Try either of these (replace the c:\msys64 as needed):

 - export QTDIR=[c:\msys64]\mingw32
 - export QTDIR=[c:\msys64]\mingw64